+++ app/Views/AdminPanel/MassEmail.php
@@ -1,7 +1,7 @@
-<div class="col-lg-8 col-md-8 col-sm-8">
+<div class='col-lg-12 col-md-12 col-sm-12'>
 	<div class="panel panel-default">
 		<div class="panel-heading">
-			<h1><?=$title;?></h1>
+			<?=$title;?>
 		</div>
 		<div class="panel-body">
 			<p><?=$welcomeMessage;?></p>